LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-page/body-margin-3a-expected.txt


FAIL Test that body's rightmargin and bottommargin attributes are ignored in standards mode assert_equals: expected "8px" but got "100px"
